Move contrib ops test
