+++ onnxruntime/test/contrib_ops/function_ops_test.cc
@@ -4,7 +4,7 @@
 #include "gtest/gtest.h"
 
 #include "core/graph/contrib_ops/contrib_defs.h"
-#include "test/providers/function_test_util.h"
+#include "test/contrib_ops/function_test_util.h"
 
 using namespace ::onnxruntime::common;
